S15.e.5b: extract operator.ts security + AGT render (1586 → 1318 LOC)
Sixth sub-slice of the S15.e operator.ts decomposition train (e.5
companion). Lifts the agent-detail render trio out of startDashboard:

cli/src/commands/operator/render/security.ts (~287 LOC):
  - renderSecurity(ctx)        — closure deps: agentTable, sandboxes,
                                 securityStates, securityBox
  - renderAGTFull(sb, sandboxes, securityStates) — pure, returns string
  - renderAGT(ctx)             — closure deps: agentTable, sandboxes,
                                 securityStates, agtPanel
  - ok(v) color-dot helper (was a top-level lambda in operator.ts)

Three thin wrappers in operator.ts preserve call sites in render().

Behavioral change: NONE.
- Bodies byte-identical to originals.
- Closure state injected via SecurityRenderContext interface.
- renderAGTFull stays pure (no widget side-effects).

Cumulative S15.e delta:
  pre-S15.e: 2894
  e.1:       2739 (-155)
  e.2:       2483 (-256)
  e.3:       1960 (-523)
  e.4:       1880  (-80)
  e.5:       1586 (-294)
  e.5b:      1318 (-268)
  Cumulative: -1576. Header + dialogs (~900 LOC) remaining for cap.

Verification:
- tsc --noEmit clean / lint 20 warnings (unchanged) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e5b.md
