+++ docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e5b.md
@@ -1,0 +1,59 @@
+# Phase 2 — S15.e.5b — operator.ts security + AGT render extraction
+
+**Date:** 2026-04-29
+**Slice:** `phase2-hotspot-operator-cli-e5b`
+**Sign-offs:** Core ✅, Security ✅
+
+## Scope
+
+Sixth sub-slice of the S15.e operator.ts decomposition train (e.5
+companion). Lifts the agent-detail render trio (`renderSecurity`,
+`renderAGTFull`, `renderAGT`) plus the `ok(v)` color-dot helper out
+of the giant `startDashboard` closure into a new render module.
+
+## What moved
+
+| File | Functions | LOC |
+|---|---|---|
+| `cli/src/commands/operator/render/security.ts` (new) | `renderSecurity(ctx)`, `renderAGTFull(sb, sandboxes, securityStates)`, `renderAGT(ctx)`, internal `ok(v)` helper | ~287 |
+
+Three thin wrappers in operator.ts so call sites in `render()` are
+unchanged.
+
+## Behavior delta
+
+**None.** Bodies byte-identical. Closure-captured `agentTable`,
+`sandboxes`, `securityStates`, `securityBox`, `agtPanel` injected via
+`SecurityRenderContext` interface. `renderAGTFull` is pure (no
+widget side-effects) so it takes positional args rather than a context
+object.
+
+The widget refs are typed structurally to avoid pulling
+`blessed-contrib` into the render modules.
+
+## LOC delta
+
+| Slice | operator.ts | Δ | Cumulative |
+|---|---|---|---|
+| pre-S15.e | 2894 | — | — |
+| S15.e.1 (#87) | 2739 | −155 | −155 |
+| S15.e.2 (#88) | 2483 | −256 | −411 |
+| S15.e.3 (#89) | 1960 | −523 | −934 |
+| S15.e.4 (#90) | 1880 | −80 | −1014 |
+| S15.e.5 (#91) | 1586 | −294 | −1308 |
+| **S15.e.5b** (this PR) | **1318** | **−268** | **−1576** |
+| §4.2 cap | 800 | | (S15.e.5c header + S15.e.6 dialogs remaining) |
+
+## Verification
+
+- ✅ `npx tsc --noEmit` clean
+- ✅ `npm run lint` 20 warnings (unchanged), 0 errors
+- ✅ `npm run build` clean
+- ✅ `npm test -- --run` → 454 pass / 2 skipped (pre-existing)
+
+## Next slices
+
+- **S15.e.5c** — `renderHeader` + `healthSummary` + `render()` orchestrator
+  extraction (~200 LOC).
+- **S15.e.6** — modal/dialog state machine (`startEdit`, `launch`,
+  `connectToAgent`, `deleteSelectedAgent`, ~700 LOC).
